requirements.txt
absl-py==2.1.0
aenum==3.1.15
affine==2.4.0
alembic==1.14.1
ansi2html==1.9.2
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
blinker==1.8.2
boto3==1.37.38
botocore==1.37.38
cachetools==5.5.1
catboost==1.2.8
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.0
colorlog==6.9.0
colour==0.1.5
comm==0.2.2
contourpy==1.1.1
creds==1.0.16
cycler==0.12.1
Cython==3.0.11
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_auth==2.3.0
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
dtreeviz==2.2.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
explainerdashboard==0.4.7
filelock==3.16.1
fiona==1.10.1
Flask==3.0.3
Flask-WTF==1.2.1
flask_simplelogin==0.2.0
fonttools==4.53.0
fsspec==2025.3.0
geojson==3.2.0
geomet==1.1.0
geopandas==0.13.2
google-auth==2.38.0
google-auth-oauthlib==1.0.0
graphviz==0.20.3
greenlet==3.1.1
h5py==3.11.0
html2text==2024.2.26
idna==3.7
imageio==2.34.2
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
IProgress==0.4
ipykernel==6.29.4
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter-dash==0.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lightgbm==4.6.0
llvmlite==0.41.1
Mako==1.3.10
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.3
optuna==4.4.0
oyaml==1.0
packaging==24.0
pandas==2.0.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
pmdarima==2.0.4
prompt_toolkit==3.0.45
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.1.2
pyproj==3.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
rasterio==1.3.11
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rsa==4.9
s3cmd==2.4.0
s3transfer==0.11.5
scikit-learn==1.3.2
scikit-posthocs==0.8.1
scipy==1.10.1
seaborn==0.13.2
sentinelhub==3.10.2
sentinelsat==1.2.1
shap==0.44.1
shapely==2.0.4
six==1.16.0
slicer==0.0.7
snuggs==1.4.7
SQLAlchemy==2.0.41
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tomli==2.0.1
tomli_w==1.0.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
tslearn==0.6.3
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2024.1
urllib3==1.26.20
utm==0.7.0
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.17.2
WTForms==3.1.2
xgboost==2.1.2
zipp==3.19.0

